+++ grako/__init__.py
@@ -3,9 +3,16 @@
 # Copyright (C) 2012-2016 by Juancarlo Añez and Thomas Bragg
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from grako._config import __version__  # noqa
-from grako._config import __toolname__  # noqa
-from grako.tool import gencode, genmodel, main  # noqa
+from grako._config import __version__
+from grako._config import __toolname__
+from grako.tool import compile, parse, gencode
+from grako.tool import main
+
+assert __version__
+assert __toolname__
+assert compile
+assert parse
+assert gencode
 
 
 if __name__ == '__main__':